Renamed util.py to dataset_config.py
* Also started applying static type-checking.
