+++ oceannavigator/views.py
@@ -9,7 +9,7 @@ from io import BytesIO
 from PIL import Image
 
 from oceannavigator import app
-from oceannavigator.util import (
+from oceannavigator.dataset_config import (
     get_variable_name, get_datasets, get_dataset_config,
     get_dataset_url, get_dataset_climatology, get_variable_scale,
     is_variable_hidden, get_dataset_cache, get_dataset_help,
