+++ plotting/stick.py
@@ -2,7 +2,7 @@ import matplotlib.pyplot as plt
 import numpy as np
 import plotting.utils as utils
 import plotting.point as plPoint
-from oceannavigator.util import get_dataset_url
+from oceannavigator.dataset_config import get_dataset_url
 from flask_babel import gettext
 from matplotlib.dates import date2num
 from data import open_dataset
