+++ plotting/transect.py
@@ -10,7 +10,7 @@ import plotting.colormap as colormap
 import plotting.utils as utils
 from oceannavigator import app
 from geopy.distance import VincentyDistance
-from oceannavigator.util import get_variable_name, get_variable_unit, \
+from oceannavigator.dataset_config import get_variable_name, get_variable_unit, \
     get_dataset_url, get_variable_scale_factor
 import plotting.line as pl
 from flask_babel import gettext
