+++ plotting/ts.py
@@ -1,6 +1,6 @@
 import matplotlib.pyplot as plt
 import numpy as np
-from oceannavigator.util import get_dataset_url, get_dataset_name, get_variable_unit
+from oceannavigator.dataset_config import get_dataset_url, get_dataset_name, get_variable_unit
 import seawater
 import plotting.point as plPoint
 import matplotlib.gridspec as gridspec
